refactor(tests): scope fixture helpers by corpus
