+++ QtProject/tests/shared/video_extraction_test_helpers.cpp
