+++ QtProject/tests/shared/video_matching_test_helpers.cpp
